LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/pseudo-elements-002-expected.txt


FAIL Check that transitions run on a pseudo element whose ancestor changes display type. assert_equals: expected "200px" but got "100px"
